pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "dinosaur"
version = "0.1.0"
authors = [
    {name = "Google LLC", email = "noreply@google.com"},
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
# TODO(shoyer): thin these down
dependencies = [
    "fsspec",
    "jax",
    "jaxlib",
    "numpy",
    "pandas",
    "pint",
    "scipy",
    "scikit-learn",
    "tree-math",
    "xarray",
    "xarray-tensorstore",
]

[project.optional-dependencies]
tests = [
    "absl-py",
    "chex",
    "pytest",
]
